Move internal project to open source
In this change, the POC of a Gerrit-slave and Gerrit-master helm chart
developed internally at SAP is transferred into an open source project.

The project contains the code to build container images used by the helm
charts and provides scripts to ease the build process.

A helm chart to deploy a Gerrit master instance is provided. Next to
Gerrit itself it provides a CronJob for Git garbage collection and
a MySQL-database.

The helm chart to deploy a Gerrit slave provides the Gerrit slave
itself, Git garbage collection, a MySQL database and a Apache-Git-based
backend to receive replication requests for repositories from a Gerrit
master.

Currently, Gerrit 2.12 is used. Both helm charts are NOT production
ready. They only represent POCs and need further development to provide
the necessary security and stability for production.

Change-Id: I913fb196af9f734bdd8c063ae5cae284d1a628d6