Enable the use of Gerrit 2.14
Gerrit versions higher than 2.13 were not yet functional, since the
database driver was not downloaded correctly. In addition the
introduction of NoteDB required to set a serverID. Gerrit can do that
automatically, but would need write access to gerrit.config, which in
the context of the helm chart is not possible.

This change takes the Gerrit-version into account, when retrieving the
MySQL-driver. For improved readability the code for retrieving the
driver was moved to a separate file.

The serverID-setting was added to the values.yaml file of the
gerrit-master and gerrit-slave helm chart and the READMEs were updated
accordingly.

Change-Id: I20ffbc1b846d0f98a1b342aa60777d11bc82b792