Force gerrit master to use gerrit.config mounted from configmap
The 'gerrit.war init' executed when starting the gerrit-master container
on Kubernetes was overwriting the symlinks to the gerrit.config and
secure.config files provided by the ConfigMap. It kept the configuration
in the process, thereby not breaking the configuration. Nevertheless
changes in the ConfigMap would not be used by Gerrit anymore.

This change sets the symlink again after the initialization, thereby
overwriting the config-files created by Gerrit.

Change-Id: If39830794a17951d0ecd64272e6954c1bab30ea5