Add cert-manager to gerrit-master chart
The cert-manager project allows to use custom resource definitions to
create certificate issuers that can provide certificates from a CA or
ACME to ingresses or other resources. This makes providing signed
certificates to the ingresses of the helm charts much easier.

This change adds the possibility to the gerrit-master chart to install
a cert-manager as a dependency of the chart and to configure certificate
issuers. The ingress regulating ingress traffic to Gerrit can be
configured to retrieve the certificates from a certificate issuer
instead of taking manually provided once. It is also possible to use
a manually installed certificate issuer.

Change-Id: I04b8c93cd1224bca46929ba1bccd8e94939645c0