Allow to publish multiple images with one command
Unlike the build-script, the publish script did not yet allow to
publish multiple/all docker images at the same time.

With this change all images may be published at the same time, thereby
speeding up the process and ensuring that all images of each version
are available in the registry.

Change-Id: I8d7df3bf6e93bcd6d34a7ec9f091f85c93b11855