Add Gerrit-version to image tags
The images were automatically tagged with the output of 'git describe
--dirty'. The version of Gerrit built into the image, which could be
configured by setting the download URL of the war-file, was not visible
without accessing the image.

With this change the build script gets the Gerrit version from the
war-file in the gerrit-base image and appends it to the version tag.

Change-Id: I4bf6fadd383fa05a21311ea1c1d74bc00ac13145