Add Gerrit-version to image tags
The images were automatically tagged with the output of 'git describe
--dirty'. The version of Gerrit built into the image, which could be
configured by setting the download URL of the war-file, was not visible
without accessing the image.

This change adds the Gerrit major.minor-version at the end of the
generated image tag, when building the script with the build-script.
The Gerrit version can be set manually using the --gerrit-version option
or the script will try to extract it from the download URL, which should
work for all URLs pointing to release builds on
http://gerrit-ci.gerritforge.com. If neither of those two ways provides
a version, 'unknown' is used as a version tag.

Change-Id: I4bf6fadd383fa05a21311ea1c1d74bc00ac13145