Describe minikube setup for evaluation
The documentation that far assumed that the user has a Kubernetes
cluster in the cloud available. For initial evaluation of the helm
charts, it would be practical to be able to test them on a local
machine. This change provides documentation and sample configuration
files to install a gerrit-master and gerrit-slave setup on a local
machine using Minikube.

Change-Id: I1eeb07401855910cca6c048e936e44c76ca6238a