Remove requirement for Java keystore in charts
The gerrit-slave and gerrit-master helm charts required a keystore to be
provided in a custom values.yaml file to work. This was actually not
necessary, when not using SSL for connections to Gerrit, which is the
default for both charts.

With this change, it is not required anymore to provide a keystore.
This allows for easier setup, when testing the charts, where SSL might
not yet be used.

Change-Id: I1d890ccec8950f574577067410cc71fb53303dec