Create ephemeral keystore in charts, when non is provided
The gerrit-slave and gerrit-master helm charts required a keystore to be
provided in a custom values.yaml file to work, since Gerrit requires
a keystore to be available.

With this change, it is not required to provide a keystore. If none is
provided, a keystore is created on pod startup. Since this keystore is
ephemeral, it will be recreated on each startup and should only be used
for testing purposes. This change makes it easier to install the charts
for evaluation and testing.

Change-Id: I1d890ccec8950f574577067410cc71fb53303dec