Turn off SSHD in gerrit.config by default
The current implementation of the gerrit-slave and gerrit-master helm
charts do not allow to expose ssh, but Gerrit was configured to start
SSHD (default for Gerrit).

This change turns off the SSHD by default in both charts. This should
make the startup faster, since the SSH daemon does not have to be
started.

Change-Id: Idc2cfdd9b97e90ff7685bd07382eb514116fd346