Provide empty All-Projects and All-Users repositories for slave
For the Gerrit slave to start, when using the gerrit-slave chart, the
All-Users.git and All-Project.git repositories were expected to exist.
The init cointainer would only proceed, if this was the case. Since it
is not necessary that these repositories already contain data for the
slave to start up, it would be easier to just create them before
starting the slave, if they do not already exist.

This change removes the parts of the init container entrypoint script
that wait for the repositories to appear and creates bare All-Projects
and All-Users repositories on startup, if they do not exit yet.

Change-Id: I1bd489d587ae530f9b8283bde6c49ccfe7f5337c