Rewrite gerrit-slave-init image to allow to handle other databases
The gerrit-slave-init container was only able to initialize using
a MySQL database. This change rewrites the scripts executed by the
container to allow to extend the functionality to other databases as
well and adds the ability to use H2 databses. Further, the container can
now also handle the initialization of a Gerrit master. Thus, the
container image was renamed to gerrit-init. The charts were adapted
accordingly. Due to these changes, this will make the charts
incompatible with previous versions.

Change-Id: Ic06de5ba6686b4c47365cbadb919c9a6ebfbd6b5