Rewrite gerrit-slave-init image to allow to handle other databases
The gerrit-slave-init container was only able to initialize a MySQL
database. The provided bash scripts weren't easy to extend to support
more databases. With Gerrit 2.16 a H2-database is usually enough, since
the content does not change during runtime. Thus, at least this database
should be supported.

This change adapts the gerrit-init container to be able to support
H2-databases. To do this several changes were needed:

- Python scripts were used instead of bash
- The gerrit-slave-init was renamed to gerrit-init, since it now also is
  able to initialize a Gerrit master out-of-the-box.
- The complete Gerrit site is now shared between containers in a pod
  using an emptyDir volume to ensure that the site initialized in
  initContainer is fully shared and not mixed with the site created at
  build time.
- The site in the emptyDir is not persisted through pod restarts. Data,
  e.g. the db-directory, that is persisted separately, overwrites the
  emptyDir.

This change makes the Gerrit initialization in the helm charts more
flexible. Also, using python and sqlalchemy to communicate with the
database is a lot faster than using bash and the mysql client, greatly
fasten up pod startup.

Change-Id: Ic06de5ba6686b4c47365cbadb919c9a6ebfbd6b5