Rewrite gerrit-init image to allow to handle other databases
The gerrit-init container was only able to initialize a MySQL
database. The provided bash scripts weren't easy to extend to support
more databases. With Gerrit 2.16 a H2-database is usually enough, since
the content does not change during runtime. Thus, at least this database
should be supported.

This change adapts the gerrit-init container to be able to support
H2-databases. To do this, python scripts were used instead of bash,
since it makes it easier to provide extensible code and since python
possesses libraries to communicate with several databases, making it
easier and more efficient to work with the databases.

This change makes the Gerrit initialization in the helm charts more
flexible. Also, using python and sqlalchemy to communicate with the
database is a lot faster than using bash and the mysql client, greatly
fasten up pod startup.

Change-Id: Ic06de5ba6686b4c47365cbadb919c9a6ebfbd6b5