Download Gerrit war-file outside of the Gerrit site
The Gerrit war-file was downloaded into the Gerrit-site. This would not
allow to persist the whole site, since volume mounts would overwrite the
site.

This change changes the download location, but sets a symlink to the old
position for compatibility reasons. A Gerrit init running on the
war-file should copy it to the site/bin folder. This way, the Gerrit
war-file should be callable, even if the site is overwritten by a volume
mount not containing it.

Change-Id: I7f5f412ac0d13444c687e23652a16e099884902a