Share the Gerrit site between containers in a pod
Only parts of the Gerrit site are persisted. The other parts were taken
from the site existing in the container. Thus, when using initContainers
the site in the initContainer and the main container could differ, which
is not ideal and could lead to unwanted effects.

This change adds an emptyDir-volume to the Gerrit-deployments in the
gerrit-master and gerrit-slave charts. This volume is shared by all
containers in the pod accessing the site. Persistent volumes used to
persist parts of the site overwrite the emptyDir volume. Some
adjustments had to be made to the startup scripts, since the emptyDir
overwrites the site present in the container on startup with an empty
directory, thereby removing resources expected by the scripts.

Change-Id: I649280f2ddf2b2d41aaf7f9953ade794ab84a900