Extract config-symlinking into function
To provide the config-files stored in ConfigMaps and maintain read-write
access to the containing directories, the files were symlinked to their
final locations instead of storing them there. This caused some code
duplication in the same scripts.

This change extracts the symlink-commands into a function to reduce code
duplication. Even if symlinking of the files did not have to be
performed multiple times in the same script, this was done for the
container-spec to stay consistent.

Change-Id: Id7066907054bfd5adb05ba670d48641726af6ae2