Move the MySQL DB into its own chart
The gerrit-master and gerrit-slave charts could be used to install
a MySQL database as a dependency to provide a relational database to
contain the ReviewDB. For Gerrit 2.16 a H2-database is usually
sufficient to store values remaining in the schema of the ReviewDB,
since they do not change during runtime. In the upcoming Gerrit-releases
the ReviewDB will be completely discontinued. Thus, no relational
database will be needed anymore. Since most people won't start to use
this charts with an older version of Gerrit, it didn't make sense to
unnecessarily inflate the charts to support old Gerrit versions.

This change removes the MySQL-database related parts from the
gerrit-master and gerrit-slave charts. To still allow an easy
installation of older Gerrit-versions, the MySQL-related parts were
moved into their own chart, called 'reviewdb', which can be installed
separately, and which works otherwise as before.

Change-Id: I9b4a1955dd645fe72c1d03bf3ecd9407622c0b1d