Add an initialization file to create a replication user
For database replication a dedicated user has to be created in the
master database. So far this had to be done manually in the reviewdb
chart.

This change adds a template for an initialization file to the
'values.yaml'-file of the reviewdb-chart that can be used to create this
user during startup of the deployment, removing this manual step.

Change-Id: Ib8b92c18b18faec4df052c70e4d6e1aef9b30cbf