Remove docker-setup-specific contents of git-gc image
The purpose of this project is to provide a Kubernetes setup for Gerrit.
Out of historical reasons the container images maintained by this
project were build in a way to also support running a similar setup
in docker. This causes a lot of additional effort and also adds contents
to the containers that are not used in the Kubernetes context, which
makes them larger and potentially more insecure than need be.

This change starts the effort of removing docker-specific content in the
docker images by removing the docker specific contents in the git-gc
image. Thus, cron was removed, the start-script was removed and sudo is
not installed by the base-image anymore, since that was only used by
cron in the meantime.

Change-Id: Ib06770227cdeee40c42ef3cd8eabd66bced00ef7