Fix Gerrit-slave initialization without test mode
In change 207458 the Gerrit site was moved into an emptyDir volume,
which is empty at pod startup. This caused the startup of the Gerrit
slave to fail, when not run in test mode, because it was not
initialized before startup.

This change adds a initContainer initializing the site in the emptyDir,
if the test mode is disabled.

Change-Id: I5339dbf37e693700292d29b354ee35c747f34109