Remove docker-setup-specific contents of apache-git-http-backend
The purpose of this project is to provide a Kubernetes setup for Gerrit.
Out of historical reasons the container images maintained by this
project were build in a way to also support running a similar setup
in docker. This causes a lot of additional effort and also adds contents
to the containers that are not used in the Kubernetes context, which
makes them larger and potentially more insecure than need be.

This change removes the start-script and the script validating the
filesystem. It removes the declaration of exposed ports in the
Dockerfile, since this is not mandatory and depending whether HTTPS or
HTTP is activated, the unused port should not be exposed. It removes the
custom entrypoint in the Kubernetes setup and moves the disabling of
unused protocol to the start-script listening to environment variables.
The fallback creation of the certificates and .htpasswd-file the start
script was removed, since due to mounting the secret into the directory
it is configured as readonly by Kubernetes. Rather the helm chart
requires the corresponding values to be set.

Change-Id: Ib9e562e08d2bf8c2f0d8136ca5e0eb13512320ca