Remove docker-setup-specific contents of gerrit-slave image
The purpose of this project is to provide a Kubernetes setup for Gerrit.
Out of historical reasons the container images maintained by this
project were build in a way to also support running a similar setup
in docker. This causes a lot of additional effort and also adds contents
to the containers that are not used in the Kubernetes context, which
makes them larger and potentially more insecure than need be.

This change moves the custom entrypoint in the gerrit-slave chart to
the start-script of the docker image. The docker container will continue
to work on its own, but has the option to symlink configuration provided
at /var/config into /var/gerrit/etc as needed for the gerrit-slave helm
chart. The script to start the container in a docker setup was removed.

Change-Id: I28166aaab8785522c2c8101f46903c564909705c