Remove docker-setup-specific contents of gerrit-init image
The purpose of this project is to provide a Kubernetes setup for Gerrit.
Out of historical reasons the container images maintained by this
project were build in a way to also support running a similar setup
in docker. This causes a lot of additional effort and also adds contents
to the containers that are not used in the Kubernetes context, which
makes them larger and potentially more insecure than need be.

This change removes the check of the filesystem permissions and the
script starting the container in the docker setup. Since the script used
to start up all containers in the docker setup does not start up any
container anymore, it is removed as well.

TODO: Wait for rewrite to be merged and adapt the entrypoint.
Change-Id: I33dfb591bbb218067b81b84ee977d0897cf1a9c4