Correctly fail in build-script
If a container image failed to build, the build-script did not always
report this. This was the case, if multiple images were built at the
same time and happened because a global variable was used to track
failures, which was not correctly reset between builds.

This change changes the method to detect return codes of the docker
build command, to detect build failures more consistently.

Change-Id: I05a71e01bfbfa544a73701951d0f23c3016c21a9