Add structure test for base-container
So far the containers created in this project weren't tested
automatically.

This change adds structure tests for the base-container making a start
in the effort of implementing automated tests. The python package
'pytest' is used to run the test and is meant to provide a framework to
implement additional tests as well. Pipenv is used to ensure
a consistent test environment.

Change-Id: I04b1bdbc5b6360402b5a629c264c5c5c882f773f