Allow to use existing container images
The tests were always rebuilding the images from scratch. The rebuilding
of the images took more time then the tests itself.

This change adds the '--tag'-parameter to the pytest command, which
takes a tag marking the existing container images. If this parameter is
provided, the tests will use existing images with this tag and build
only missing images. If the parameter is not provided, all images will
be built as before.

Change-Id: I262cf58129f23c898d2efb47d14c208ee1860763