Add option to pytest to use container build cache
When building container images during testing the build cache of Docker
was disabled, enforcing a build from scratch. While this is a good
behavior for tests, this also increased testing times. During
development fast test times may be preferable.

This change adds the '--build-cache'-flag to the 'pytest'-command, which
activates usage of the build cache.

Change-Id: Ifc9bcd4ca9b49ea1ad50199734de5dbe56f91b84