Add integration tests for mysql-replication-init
The tests implemented in this change will run incrementally and test the
execution of the entrypoint script in the mysql-replication-init
container. To do this a container running a mysql-database is started
for the time of the tests. To allow communication between containers
during testing pytest will now create a docker network for the time of
the test-session, which can be used by containers to communicate.

Change-Id: I196b954d8318b215d58b5b3c0c922c917e885f0f