Bump Gerrit version to 3.0.1 and remove ReviewDB-support
This change bumps the default Gerrit-version to 3.0.1. Since this
version does not need a ReviewDB anymore and since it is unlikely that
the Kubernetes setup will be used with an earlier version of the chart,
the support for a ReviewDB is also removed with this change. This also
reduces the container size and startup time.

Change-Id: I11f3d8717f045f0b71c7b5d759a0be256d4626aa