Add integration tests for gerrit-slave container
This change adds test, that test whether the container slave correctly
starts up. Most tests are implemented in a scenario using a H2-database,
in which it is tested, whether...

- Gerrit starts successfully
- externally supplied configuration is available in the correct location
- the All-Projects and All-Users repositories are created
- cloning works using a configured port
- the UI is not available

In a second scenario, it is tested, whether the MySQL driver is downloaded,
if MySQL is configured to be used as a database.

Change-Id: Ic3992115482625635532e154c80717def576a1e4