Add integration tests for gerrit-init container with H2-database
This change adds integration tests for the gerrit-init container, that
test its behaviour, if using a H2-database to store Gerrit's reviewdb.
This test scenario is also used to test whether the gerrit-init
container correctly installs Gerrit plugins and in general allows
testing of functionality, which is not dependent on the database used by
Gerrit.

Change-Id: I6bf786dfb186c1ab194b7c9601fa189a101cbfdf