Add a helper function for tests to wait for event
A pattern used by many tests was to wait for a certain event happening
in the container and to stay in a time limit. This was implemented in
each test case using this pattern.

This change adds a helper function available to all tests, that takes
a function and a time limit. The function can return some output, which
is forwarded to the test and has to return a boolean, indicating whether
the event, for which the function tests has happened. The helper
function will execute the given function as long as this function returns
False or until the time limit is reached and will then return the
functions output and a boolean indicating, whether the function returned
True within the time limit.

Change-Id: I3505e0f25028a81f84dc7e5b8f006e0d6a904ab4