Improve the central README
The README.md file in the root directory of the project was not
up-to-date with the latest changes to the project. It still provided
instructions on how to install Gerrit on Docker using the provided
images, which is now out of scope of the project and might confuse
users.

This change removes the instructions on how to run the provided
container images using docker. It moves the introductory information on
how to deploy the helm-charts on Kubernetes to the top of the document,
since this represents the most important information for most users. The
change also adds sections of how to contribute to the project and how to
contact the Gerrit mailing list.

Change-Id: I85392eb3370a2896d3af92508a57a881429b0fad