Add pylint configuration and lint python scripts accordingly
Python files were not linted so far. Since in the following changes
tests for the container images and helm charts will be added, which will
use python, a linter would be helpful to keep the code quality high.

This change adds a pylint configuration file (.pylintrc) to the project
configuring pylint. The default configuration is changed in the folowing
ways:

- Use 2-space indentations
- Do not enforce doc-strings

The python scripts used in the gerrit-init container did not fully
follow PEP8-rules and were adapted according to the output of pylint.

Change-Id: I495ebc6b9cf47e63e40cf1b19dad6fbbbd3bc260