Provide a factory providing gerrit container
So far the provisioning of a gerrit container with a factory was
only performed in the gerrit-slave container tests. Such a
container will also be needed to test the functionality of the
gerrit-init container.

This change moves the factory for a gerrit container out of the
gerrit-slave container tests to the conftests.py file for
container-image tests. Thereby, the factory is available for all
container-image tests.

Change-Id: Ica537347bef3ba72f0de408a81c5c226aaaee14d