Extract testing for inherited image to helper function
For most container images a test existed that checked, whether the
expected image was inherited to ensure that expected base components
were available. The code checking this was identical in all tests.

This change moves the logic checking for the inherited images to
a helper function, thereby reducing code duplication.

Change-Id: I3cce7c026988c852441379960c3ac11f52811ae6