Add test setup for helm chart tests
This change implements fixtures required to set up a Kubernetes cluster
to run integrations tests of helm charts on. The tests require to be
given a kubeconfig file of an existing Kubernetes cluster. The test
fixtures will then initialize helm, a read-write-many filesystem
provisioner, an image-pull-secret and delete all created resources after
finishing the tests. This change further adds a single test to test
whether the gerrit-master helm-chart successfully deploys. This test
shows that the fixtures function correctly.

Change-Id: I106f65395f82286d86ee73198f5c996d94fbc610