Allow tests to push docker images to registry
The images built during test execution were not pushed to a registry.
This would be convenient in some scenarios and also required for testing
on Kubernetes, since the images have to be accessible over the internet
to be used by Kubernetes.

This change adds the option to push the images to a configured registry,
if the '--push'-flag is set. The registry can be set with the
'--registry'-option. Login-data for the registry may be provided using
the 'registry-user'- and 'registry-pwd'-options.

Change-Id: I9daa7b9a6c17ddd6a3c0ac4de79b7677a0ea78b7