Add readiness and liveness probes for gerrit-master
The kubelet did not have a way to see, whether Gerrit was running in the
gerrit-master pods created by the gerrit-master helm chart. Thus, the
readiness of the pod was reported to early, since by default just the
start of the entrypoint script is used and Gerrit startup took a few
seconds longer. Further, if Gerrit crashed, the pod would not restart,
since 'tail' was the process running in the foreground and thus used by
default to check the container health.

This change adds a readiness- and liveness probe to the gerrit-master
container in the respective pod and chart. These probes check for
a successful HTTP-connection on port 8080 of the pod, which is used to
connect to Gerrit. The initial delay and probe period can be configured
in the chart.

Change-Id: Ia2e13dbe809a4ab3168729614efb8ff064abfd2e