Add readiness and liveness probe for git-backend in gerrit-replica chart
This change adds a readiness- and a liveness probe to the
apache-git-http-backend containers in the respective pods. These probes
check, whether the webserver is still reachable under the '/' route,
which just serves the apache default webpage. With this change, the
apache-git-http-backend pod will wait for apache to be ready and
restart, if apache crashed.

Change-Id: I8dfafccbcdf5759eff9c9b79bc937532ce909187