Deny all ingress and egress traffic to git-gc pods
By default the ingress and egress traffic of pods is not restricted in
Kubernetes. This can be a security risk. The git-gc pods started by the
git-gc CronJob do not require any network access, which was nevertheless
not blocked.

This change adds a NetworkPolicy for the git-gc pods in the
gerrit-master and gerrit-slave charts that blocks all ingress and
egress traffic.

Change-Id: Ie8ea090e8b39f31f1d64e2916069bee9e0bb6994