Do not reindex Lucene indices on every start of the gerrit-master pod
The gerrit-master pod was reindexing all Gerrit indices on each startup.
This could be expensive, if the amount of data in the indices would be
large.

This change moves the reindexing to the gerrit-init container.
A python-script was added, that runs reindexing, if indices are not
ready. This is the case in the following cases:

- The $gerrit_site/index/gerrit_index.config does not exist
- At least one of the indices is marked as not ready in the
  gerrit_index.config file.
- The index versions are different between the Lucene indices and the
  gerrit_index.config

Further if the chart is configured to not persist the Gerrit site,
reindexing will be enforced.

Change-Id: I65e74f6b2d55788737fa5b67b61b1d25e6ebb925