Do not reindex Lucene indices on every start of the gerrit-master pod
The gerrit-master pod was reindexing all Gerrit indices on each startup.
This could be expensive, if indices were quite large.

This change moves the reindexing to the gerrit-init container.
A python-script was added, that run reindexing, if indices are not
ready. This is the case n the following cases:

- The $gerrit_site/index/gerrit_index.config does not exist
- At least one one of the indices is marked as not ready in the
  gerrit_index.config file.
- The index versions are different between the Lucene indices and the
  gerrit_index.config

Further if the chart is configured to not persist Lucene indices,
reindexing will be enforced.

Checking Elasticsearch indices was not implemented so far. Thus,
Elasticsearch indices will still be reindexed on every pod restart.

Change-Id: I65e74f6b2d55788737fa5b67b61b1d25e6ebb925