Add NetworkPolicies to the reviewdb chart
By default, the network traffic to and from Kubernetes pods is not
restricted, which poses a potential security risk. The traffic can be
restricted using the NetworkPolicy resource of Kubernetes, which hadn't
been done so far for the reviewdb chart.

This change adds basic NetworkPolicies to the gerrit-slave chart:

- All traffic is blocked by default to and from pods installed by the
  chart.
- Egress to DNS services is allowed. The required ports can be
  configured in the chart.
- Egress and Ingress to the database pods can be configured by adding
  podSelectors, namespaceSelectors or ipBlocks. Additional ports may not
  be opened (only container port 3306 is configured).
- NetworkPolicies require pod labels to select pods, which they affect.
  The mysql-chart allows to set custom labels, which will then also used
  for the NetworkPolicies.

Using the NetworkPolicies of the chart is optional and disabled by
default to reduce the initial complexity.

Change-Id: Ibc8d4eb92d87d2c10609b30d4fcfb01a7b269c9a