Update MySQL-chart dependency to 0.15.0 for reviewdb chart
The MySQL-chart version used for the reviewdb-chart was behind a few
versions. The previously used version also did not allow to set custom
labels to the database pods, which thus could not be selected by other
resources.

This change updates the chart to the newest version, which is also
a requirement for NetworkPolicies to work, since they require pod labels.

Change-Id: I5dacac5c180d3513d9cf083d48c36d605e50b0fc