Allow to select packaged plugins in gerrit-master chart
The gerrit-master chart installed some selected plugins packaged in the
gerrit.war on startup, but did not allow to select which core plugins
should be installed.

This change adds an option to the values.yaml chart, which allows to
hand over a list of packaged plugins to install for Gerrit. The default
value is the fixed list of plugins, which was previously installed:

    - commit-message-length-validator
    - download-commands
    - replication
    - reviewnotes

Change-Id: I88900e62824f27e51da72387dc7630b360cb5da8