Allow to install custom Gerrit-plugins in gerrit-master chart
It was not yet possible to install plugins that were not packaged plugins in
the gerrit-master chart.

This change allows to configure a list of plugins to install for Gerrit.
To do this, it is required to give a URL, where the plugin jar can be
downloaded. Additionally, a SHA1-checksum has to be provided to ensure that
the same jar is used for each pod. The plugins are downloaded in the
`gerrit-init` init-container before Gerrit starts up, as long as the jar
is not yet present in the plugins directory. Optionally, the plugin
files can be cached.

Change-Id: I34e6bc5668090b37c9749d4df72eeb719fdf3807