Allow to install custom Gerrit-plugins in gerrit-master chart
It was not yet possible to install plugins that were not core plugins in
the gerrit-master chart.

This change allows to configure a list of plugins to install for Gerrit.
To do this, it is required to give a URL, where the plugin jar can be
downloaded. Optionally, a SHA256-checksum can be provided to ensure that
the same jar is used for each pod. The plugins are downloaded in an
init-container before Gerrit starts up, as long as the jar is not yet
present in the plugins directory.

Change-Id: I34e6bc5668090b37c9749d4df72eeb719fdf3807