Use a StatefulSet to manage Gerrit pods in the gerrit-master chart
So far a Deployment was used to manage Gerrit pods in the gerrit-master
chart. This did not give the pods a stable, discoverable address and
stable volume association, which would be required to use multiple
Gerrit masters, e.g. with the HA-plugin.

With this change a StatefulSet is used instead to manage the
Gerrit-pods. The StatefulSet provides each pod with a stable name and
stable RWO volumes.

So far only one pod replica is fully supported, since Gerrit is not
configured for a multi-master setup.

The RWM-volume used to store git repositories is not yet stable, since
StatefulSets are not able to manage volumes that are shared between pods
of the StatefulSet. Thus, this volume will not survive chart-deletion
and there is no guarantee that the exact volume is reattached during
a pod restart.

Change-Id: I814d1be16a0723b01d731525f8185f299d38b3ea