Add option to provide external PVC for git repositories
If the helm chart was deleted, the volume containing the git
repositories was removed and could not be mounted anymore into a Gerrit
pod, if the chart was reinstalled.

This change provides an option to provide an external PVC for the volume
containing git repositories. Since this PVC is not managed by helm, it
will not be deleted with the chart and can be mounted again. This
feature is optional and disabled by default, meaning that the helm chart
will create its own PVC by default.

Change-Id: I3bb966095970caf96be65bfb8c05faf83500b795