Use symlinks to bring volume mounts into site
Mounted volumes containing parts of the Gerrit site were currently
mounted on top of the empty site containing the rest of the Gerrit site.
This could potentially lead to issues, if the order of mounting the
volumes was incorrect.

This change uses symlinks instead. The volumes are mounted to
a different location in the container (var/mnt/$dir) and then symlinked
to the gerrit site.

Change-Id: If03bf2e19f772ec95184a0d09b510c8aefced1b6