Persist the Gerrit site in a single volume
So far most of the Gerrit site was stored in an EmptyDir-Volume and only
some parts were each persisted in a dedicated persistent volume. While
providing flexibility this approach would lead to a huge amount of
persistent volumes and could easily cause inconsistencies.

Now the whole Gerrit site except for the repositories are stored in
a single persistent volume. This was changed for the gerrit-master and
the gerrit-slave chart.

Change-Id: I08ebd8d0166363ddffd0e182575d900649780ac1